refactor(whispering): extract presentational RecordingPill component
Split the recording overlay route into a platform-free RecordingPill.svelte
(props status/level, callback props onStop/onCancel/onFocusMain, zero Tauri
imports) and a thin Tauri adapter route that keeps the listen/emit IPC glue and
the document-level overlay-webview styles. No behavior change: the pill still
renders only the RECORDING and VAD states. Sets up sharing the same component
on web.
